+++ Installer/OS/dragonOS_focal.py
@@ -2368,6 +2368,12 @@ programs_dragonOS_noble.append(('IQ Flow Graphs',
 find . -name '*.grc' -exec grcc {} \;
 """,True,'Compile Flow Graphs'))
 
+# Plugin Flow Graphs
+programs_dragonOS_noble.append(('Plugin Flow Graphs',
+"""cd """ + fissure_directory + """/Plugins/
+find . -name '*.grc' -exec grcc {} \;
+""",True,'Compile Flow Graphs'))
+
 # PD Flow Graphs
 programs_dragonOS_noble.append(('PD Flow Graphs',
 """cd """ + fissure_directory + """/Flow\ Graph\ Library/maint-3.10/PD\ Flow\ Graphs/