+++ Installer/OS/ubuntu22_04.py
@@ -2317,6 +2317,12 @@ programs_ubuntu22_04.append(('IQ Flow Graphs',
 find . -name '*.grc' -exec grcc {} \;
 """,True,'Compile Flow Graphs'))
 
+# Plugin Flow Graphs
+programs_ubuntu22_04.append(('Plugin Flow Graphs',
+"""cd """ + fissure_directory + """/Plugins/
+find . -name '*.grc' -exec grcc {} \;
+""",True,'Compile Flow Graphs'))
+
 # PD Flow Graphs
 programs_ubuntu22_04.append(('PD Flow Graphs',
 """cd """ + fissure_directory + """/Flow\ Graph\ Library/maint-3.10/PD\ Flow\ Graphs/